Simplify significantly by removing a layer of generic-ness
Instead of having data-defined categories, the categories are now going to just be defined as explicit model objects (currently there's only one: Game).

The model object will have a module that has various HTTP handlers and support classes for it.

To avoid a lot of boilerplate code, there is now an Items.Helpers module which contains a lot of functionality for reflexively converting between model objects, JObjects, and form data fields.

This should make it faster to build out new category types going forward (Currently, there's only about 130 lines of Game-specific code; most of that will need to be copy-pasted into new objects which is a bummer but it's far less confusing than the previous scheme -- this will grow a bit once I figure out image uploading, but that's for another commit)