Fixed an issue where the microblogs were not appearing on add/edit views
This was happening because the pageData was
not being set because we weren't correctly
finding the id field.

This is because the "view" fields were being
passed in ot addEditView and the "view" fields
don't contain the id field (since the id is not in
the view).

Now we just pass the "allFields" in everywhere
and the method that's receiving it can filter it
down if appropriate